move component to its place
