+++ src/main/java/kaptainwutax/minemap/feature/chests/loot/BurriedTreasureLoot.java
@@ -24,5 +24,4 @@ public class BurriedTreasureLoot extends Loot {
         return Collections.singletonList(loot1);
     }
 
-
 }
